+++ packages/services/src/Domain/Device/MobileDeviceInterface.ts
@@ -1,5 +1,7 @@
+import { AppleIAPProductId } from './../Subscription/AppleIAPProductId'
 import { DeviceInterface } from './DeviceInterface'
 import { Environment, Platform, RawKeychainValue } from '@standardnotes/models'
+import { AppleIAPReceipt } from '../Subscription/AppleIAPReceipt'
 
 export interface MobileDeviceInterface extends DeviceInterface {
   environment: Environment.Mobile
@@ -22,4 +24,5 @@ export interface MobileDeviceInterface extends DeviceInterface {
   isUrlComponentUrl(url: string): boolean
   getAppState(): Promise<'active' | 'background' | 'inactive' | 'unknown' | 'extension'>
   getColorScheme(): Promise<'light' | 'dark' | null | undefined>
+  purchaseSubscriptionIAP(plan: AppleIAPProductId): Promise<AppleIAPReceipt | undefined>
 }