+++ packages/services/src/Domain/Feature/FeaturesEvent.ts
@@ -1,4 +1,5 @@
 export enum FeaturesEvent {
   UserRolesChanged = 'UserRolesChanged',
   FeaturesUpdated = 'FeaturesUpdated',
+  DidPurchaseSubscription = 'DidPurchaseSubscription',
 }