+++ packages/services/src/Domain/Subscription/AppleIAPReceipt.ts
@@ -1,0 +1,8 @@
+import { AppleIAPProductId } from './AppleIAPProductId'
+
+export type AppleIAPReceipt = {
+  productId: AppleIAPProductId
+  transactionDate: string
+  transactionId: string
+  transactionReceipt: string
+}