+++ packages/services/src/Domain/Subscription/SubscriptionManager.ts
@@ -4,6 +4,7 @@ import { InternalEventBusInterface } from '../Internal/InternalEventBusInterface
 import { AbstractService } from '../Service/AbstractService'
 import { SubscriptionClientInterface } from './SubscriptionClientInterface'
 import { Uuid } from '@standardnotes/common'
+import { AppleIAPReceipt } from './AppleIAPReceipt'
 
 export class SubscriptionManager extends AbstractService implements SubscriptionClientInterface {
   constructor(
@@ -54,6 +55,26 @@ export class SubscriptionManager extends AbstractService implements Subscription
       return result.data.success === true
     } catch (error) {
       return false
+    }
+  }
+
+  async confirmAppleIAP(
+    params: AppleIAPReceipt,
+    subscriptionToken: string,
+  ): Promise<{ success: true } | { success: false; message: string }> {
+    try {
+      const result = await this.subscriptionApiService.confirmAppleIAP({
+        ...params,
+        subscription_token: subscriptionToken,
+      })
+
+      if (result.data.error) {
+        return { success: false, message: result.data.error.message }
+      }
+
+      return result.data
+    } catch (error) {
+      return { success: false, message: 'Could not confirm IAP.' }
     }
   }
 }