+++ packages/web/src/javascripts/Application/Application.ts
@@ -39,6 +39,7 @@ import { AndroidBackHandler } from '@/NativeMobileWeb/AndroidBackHandler'
 import { PrefDefaults } from '@/Constants/PrefDefaults'
 import { setCustomViewportHeight } from '@/setViewportHeightWithFallback'
 import { WebServices } from './WebServices'
+import { FeatureName } from '@/Controllers/FeatureName'
 
 export type WebEventObserver = (event: WebAppEvent, data?: unknown) => void
 
@@ -204,10 +205,6 @@ export class WebApplication extends SNApplication implements WebApplicationInter
     return this.isNativeMobileWeb() && this.platform === Platform.Ios
   }
 
-  get hideSubscriptionMarketing() {
-    return this.isNativeIOS()
-  }
-
   mobileDevice(): MobileDeviceInterface {
     if (!this.isNativeMobileWeb()) {
       throw Error('Attempting to access device as mobile device on non mobile platform')
@@ -349,6 +346,14 @@ export class WebApplication extends SNApplication implements WebApplicationInter
 
   entitledToPerTagPreferences(): boolean {
     return this.hasValidSubscription()
+  }
+
+  get entitledToFiles(): boolean {
+    return this.getViewControllerManager().featuresController.entitledToFiles
+  }
+
+  showPremiumModal(featureName: FeatureName): void {
+    void this.getViewControllerManager().featuresController.showPremiumAlert(featureName)
   }
 
   hasValidSubscription(): boolean {