+++ packages/web/src/javascripts/Components/ContentListView/Header/ContentListHeader.tsx
@@ -153,7 +153,7 @@ const ContentListHeader = ({
   }, [OptionsMenu, AddButton, FolderName])
 
   return (
-    <div className="section-title-bar-header items-start gap-1 overflow-hidden">
+    <div className="section-title-bar-header items-start gap-1">
       {!isTablet && PhoneAndDesktopLayout}
       {isTablet && TabletLayout}
     </div>