+++ packages/web/src/javascripts/Components/PremiumFeaturesModal/PremiumFeatureModalType.tsx
@@ -1,0 +1,4 @@
+export enum PremiumFeatureModalType {
+  UpgradePrompt,
+  UpgradeSuccess,
+}