+++ packages/web/src/javascripts/Controllers/FeatureName.ts
@@ -1,0 +1,3 @@
+export enum FeatureName {
+  Files = 'Encrypted File Storage',
+}