+++ packages/web/src/javascripts/Controllers/FeaturesController.ts
@@ -1,4 +1,6 @@
+import { FeatureName } from './FeatureName'
 import { WebApplication } from '@/Application/Application'
+import { PremiumFeatureModalType } from '@/Components/PremiumFeaturesModal/PremiumFeatureModalType'
 import { destroyAllObjectProperties } from '@/Utils'
 import {
   ApplicationEvent,
@@ -14,8 +16,9 @@ import { CrossControllerEvent } from './CrossControllerEvent'
 export class FeaturesController extends AbstractViewController {
   hasFolders: boolean
   hasSmartViews: boolean
-  hasFiles: boolean
+  entitledToFiles: boolean
   premiumAlertFeatureName: string | undefined
+  premiumAlertType: PremiumFeatureModalType | undefined = undefined
 
   override deinit() {
     super.deinit()
@@ -23,8 +26,9 @@ export class FeaturesController extends AbstractViewController {
     ;(this.closePremiumAlert as unknown) = undefined
     ;(this.hasFolders as unknown) = undefined
     ;(this.hasSmartViews as unknown) = undefined
-    ;(this.hasFiles as unknown) = undefined
+    ;(this.entitledToFiles as unknown) = undefined
     ;(this.premiumAlertFeatureName as unknown) = undefined
+    ;(this.premiumAlertType as unknown) = undefined
 
     destroyAllObjectProperties(this)
   }
@@ -34,7 +38,7 @@ export class FeaturesController extends AbstractViewController {
 
     this.hasFolders = this.isEntitledToFolders()
     this.hasSmartViews = this.isEntitledToSmartViews()
-    this.hasFiles = this.isEntitledToFiles()
+    this.entitledToFiles = this.isEntitledToFiles()
     this.premiumAlertFeatureName = undefined
 
     eventBus.addEventHandler(this, CrossControllerEvent.DisplayPremiumModal)
@@ -42,11 +46,12 @@ export class FeaturesController extends AbstractViewController {
     makeObservable(this, {
       hasFolders: observable,
       hasSmartViews: observable,
-      hasFiles: observable,
-
+      entitledToFiles: observable,
+      premiumAlertType: observable,
       premiumAlertFeatureName: observable,
       showPremiumAlert: action,
       closePremiumAlert: action,
+      showPurchaseSuccessAlert: action,
     })
 
     this.showPremiumAlert = this.showPremiumAlert.bind(this)
@@ -55,12 +60,15 @@ export class FeaturesController extends AbstractViewController {
     this.disposers.push(
       application.addEventObserver(async (event) => {
         switch (event) {
+          case ApplicationEvent.DidPurchaseSubscription:
+            this.showPurchaseSuccessAlert()
+            break
           case ApplicationEvent.FeaturesUpdated:
           case ApplicationEvent.Launched:
             runInAction(() => {
               this.hasFolders = this.isEntitledToFolders()
               this.hasSmartViews = this.isEntitledToSmartViews()
-              this.hasFiles = this.isEntitledToFiles()
+              this.entitledToFiles = this.isEntitledToFiles()
             })
         }
       }),
@@ -74,13 +82,19 @@ export class FeaturesController extends AbstractViewController {
     }
   }
 
-  public async showPremiumAlert(featureName: string): Promise<void> {
+  public async showPremiumAlert(featureName: FeatureName | string): Promise<void> {
     this.premiumAlertFeatureName = featureName
-    return when(() => this.premiumAlertFeatureName === undefined)
+    this.premiumAlertType = PremiumFeatureModalType.UpgradePrompt
+
+    return when(() => this.premiumAlertType === undefined)
+  }
+
+  showPurchaseSuccessAlert = () => {
+    this.premiumAlertType = PremiumFeatureModalType.UpgradeSuccess
   }
 
   public closePremiumAlert() {
-    this.premiumAlertFeatureName = undefined
+    this.premiumAlertType = undefined
   }
 
   private isEntitledToFiles(): boolean {