+++ packages/releases/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.227](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.226...@standardnotes/releases@1.3.227) (2022-11-13)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.226](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.225...@standardnotes/releases@1.3.226) (2022-11-13)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.225](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.224...@standardnotes/releases@1.3.225) (2022-11-13)
+
+**Note:** Version bump only for package @standardnotes/releases
+
 ## [1.3.224](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.223...@standardnotes/releases@1.3.224) (2022-11-11)
 
 **Note:** Version bump only for package @standardnotes/releases